added scroll features
added features on the form pages so it automatically scrolls when the page is opened, the index page just has a link that can be styled as a button. That page does not auto scroll. Also added CSS to be able to hide the <a> used by the javascript to scroll